STL files
Use the "LCD-no-buttons" if the rotary encoder button works for you. Otherwise, use the "LCD-button" but you'll have to get a button and modify the encoder slightly. "Clip" is only needed if "LCD-button" is chosen. "Tubing-spacer" 1 and 2 are just 2 different versions of the pump hardware to account for variations in thickness. An alternative pump head design is available if needed, but this should fit most needs using 1 mm thick sylastic tubing.